+++ Source/Core/Core/Slippi/SlippiDirectCodes.h
@@ -25,7 +25,7 @@ class SlippiDirectCodes
 
 	void ReadFile();
 	void AddOrUpdateCode(std::string code);
-	std::string get(u8 index);
+	std::string get(int index);
 	void Sort(u8 sortByProperty = SlippiDirectCodes::SORT_BY_TIME);
 	std::string Autocomplete(std::string startText);
